Separar API linea de comandos y OpenBadges
