+++ slack_badges_bot/adapters/__init__.py
@@ -1,1 +1,1 @@
-from . import api, repositories, slack
+from . import api, repositories, slack, openbadgesapi
