+++ slack_badges_bot/services/__init__.py
@@ -1,1 +1,1 @@
-#from . import badge, config, repositories, award, person, issuer
+from . import badge, config, repositories, award, person, issuer
